# 11_OPENAPI.yaml
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": true,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "APIError",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Preview": {
        "additionalProperties": true,
        "properties": {
          "collected_at": {
            "title": "Collected At",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "report": {
            "additionalProperties": true,
            "title": "Report",
            "type": "object"
          }
        },
        "required": [
          "kind",
          "collected_at",
          "count",
          "items",
          "report"
        ],
        "title": "Preview",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "IMPLEMENTED local single-user API. No authentication is enforced. Loopback is the default deployment boundary, not an access-control guarantee. U06 privacy/access policy remains planned. Canonical 11_OPENAPI.yaml is generated from this app; /api is preserved. No /api/v1 aliases or target-only resources.",
    "title": "VK Social Radar runtime API",
    "version": "0.4.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/changes": {
      "get": {
        "description": "Membership timeline only; new events include immutable snapshot pair provenance, legacy events labelled legacy_unknown. Limit clamps to 1..500; no rejected range.",
        "operationId": "get_changes",
        "parameters": [
          {
            "description": "Clamped to 1..500; out-of-range integers are accepted",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Clamped to 1..500; out-of-range integers are accepted",
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Change"
                  },
                  "title": "Response Get Changes",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Get Changes",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/api/collector/check-auth": {
      "post": {
        "description": "Checks VK login inside the separate browser; this is not API authentication.",
        "operationId": "collector_check_auth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectorStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Domain/import/collector input or operation failure."
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Check Auth",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/api/collector/close": {
      "post": {
        "description": "Closes the collector browser and returns status.",
        "operationId": "collector_close",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectorStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Close",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/api/collector/collect/{kind}": {
      "post": {
        "description": "Supported kind friends/followers/dialogs; invalid kind, busy, stopped, unauthenticated or zero valid items maps 400. Nonempty observation is a preview, not a complete snapshot. No automatic save.",
        "operationId": "collector_collect",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preview"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Domain/import/collector input or operation failure."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Collect",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/api/collector/navigate/{target}": {
      "post": {
        "description": "Supported target values home/friends/followers/dialogs; invalid target or collector failure is 400, not framework enum validation.",
        "operationId": "collector_navigate",
        "parameters": [
          {
            "in": "path",
            "name": "target",
            "required": true,
            "schema": {
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectorStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Domain/import/collector input or operation failure."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Navigate",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/api/collector/organization-source": {
      "post": {
        "description": "Synchronous public organization preview with extensible report/diagnostics; result may describe partial/failure in a 200 body. Generic save-preview does not support organization previews.",
        "operationId": "collector_organization_source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "description": "source_url is string-coerced and stripped; absent/falsy/blank is 400. options is passed through only if an object; otherwise treated as {}. Collector restrictions are business errors, not framework enums.",
                "properties": {
                  "options": {
                    "description": "Only objects are used. Synchronous collection int-coerces and clamps: max_member_index or max_profiles (default 7000, 1..20000), max_pages or max_scrolls (800, 1..2000), max_candidate_profile_enrichment (30, 0..200), max_ambiguous_profile_enrichment (10, 0..100), timeout_ms (30000, 5000..60000). Falsy values select defaults; malformed values yield 400 synchronously or job failure asynchronously. Unknown keys ignored. These are coercion rules, not framework request bounds."
                  },
                  "source_url": {}
                },
                "required": [
                  "source_url"
                ],
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Domain/import/collector input or operation failure."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Organization Source",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
...
    "/api/collector/preview": {
      "get": {
        "description": "Last in-memory relation/dialog/organization preview; missing preview 404. Items/report retain collector-specific diagnostic fields.",
        "operationId": "collector_preview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Preview"
                    },
                    {
                      "$ref": "#/components/schemas/OrganizationResult"
                    }
                  ],
                  "title": "Response Collector Preview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Person, preview or collector operation not found."
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Unhandled server failure (default plain text)."
          }
        },
        "summary": "Collector Preview",
        "x-audience": "public",
        "x-implementation-status": "IMPLEMENTED"
      }
    },
...
  },
  "servers": [
    {
      "description": "Default local loopback server",
      "url": "http://127.0.0.1:8765"
    }
  ]
}